reorganize views/projects/ subfolders
