+++ moped-editor/src/utils/mapHelpers.js
@@ -1,5 +1,5 @@
 import React from "react";
-import theme from "../../theme/index";
+import theme from "../theme/index";
 
 export const MAPBOX_TOKEN = `pk.eyJ1Ijoiam9obmNsYXJ5IiwiYSI6ImNraWV4dHR0ZjAwNnYyd3FwYjFoNHduaDcifQ.--3vRm2KHq1gh5K_L0pqtA`;
